requirements.txt
Django==4.1.13
djangorestframework
djangorestframework-jwt
django-cors-headers
djangorestframework-simplejwt